fix: clean insert
